+++ Makefile
@@ -7,13 +7,15 @@ CXX := c++
 CXXFLAGS = -Wall -Wextra -Werror -std=c++98 -pedantic
 
 CXXFLAGS += -I src
-CXXFLAGS += -MMD
+CXXFLAGS += -MMD -MP
 
-SUBDIRS := strutil util logger color general
+SRC_EVENT := \
+	$(wildcard src/event/*.cpp) \
+	$(foreach dir, kqueue, $(wildcard src/event/$(dir)/*.cpp))
+SRC_UTIL := $(foreach dir, strutil util logger color general, $(wildcard src/util/$(dir)/*.cpp))
+SRC_SOCKET := $(wildcard src/socket/*.cpp)
 
-SRC_UTIL := $(foreach dir, $(SUBDIRS), $(wildcard src/util/$(dir)/*.cpp))
-
-SRC := src/main.cpp $(SRC_UTIL)
+SRC := src/main.cpp $(SRC_UTIL) $(SRC_EVENT) $(SRC_SOCKET) $(wildcard src/*.cpp)
 OBJ := $(SRC:%.cpp=%.o)
 
 $(NAME): $(OBJ)
